Implement hourly-derived day/night rain chances for non-NWS APIs
Summary of Changes:
- **Core Implementation (ForecastRepository.kt)**:
    - Enhanced mapDailyForecast to extract Daytime (8 AM - 8 PM) and Nighttime (8 PM - 8 AM) precipitation probabilities from the hourly data stream for all weather providers.
    - Updated fetch orchestration for Open-Meteo, Tomorrow.io, WeatherAPI, Visual Crossing, Silurian, and OpenWeatherMap to pass hourly data into this mapping logic.
- **UI Logic (DailyViewLogic.kt)**:
    - Refactored buildDailyRainLabel to use the period-specific daytime maximum instead of falling back to the daily summary maximum, preventing nighttime rain spikes from inflating daytime labels.
    - Enabled nighttime rain labels for 'Today' (Tonight) by calculating period-specific maximums from the hourly stream.
- **Unified Windows (DailyForecastIconResolver.kt)**:
    - Standardized on fixed 8 AM - 8 PM and 8 PM - 8 AM windows for all sources, replacing legacy dynamic sun-time logic for consistent period-specific calculations.
- **Verification**:
    - Created ForecastRepositoryDayNightPrecipTest to verify window bucketing and aggregation.
    - Added OpenMeteoDayNightPrecipIntegrationTest to verify the full fetch-to-database flow.
    - Added regression tests in DailyViewLogicTest confirming that daytime labels prefer daytime-specific data and that 'Tonight' labels appear for all sources.

Verification:
- Build: Successful.
- Unit Tests: All 52 tests in DailyViewLogicTest passed.
- Integration: Verified correct database population on emulator for Silurian tomorrow (10% day, 50% night).

+++ app/src/main/java/com/weatherwidget/data/repository/ForecastRepository.kt
@@ -350,13 +350,14 @@ class ForecastRepository
                     mapDailyForecast(
                         DailyForecast(
                             day.date,
-                            day.highTemp.toFloat(),
-                            day.lowTemp.toFloat(),
+                            day.highTemp,
+                            day.lowTemp,
                             day.condition,
                             day.condition,
                             day.precipProbability,
                             day.precipAmountMm,
                         ),
+
                         latitude,
                         longitude,
                         locationName,
